[PLAT-760] Extract status concern
