+++ lib/background_worker/base.rb
@@ -1,47 +1,24 @@
 require 'background_worker/persistent_state'
 require 'background_worker/worker_execution'
 require 'background_worker/logging'
+require 'background_worker/state'
 
 module BackgroundWorker
   class Base
     include BackgroundWorker::Logging
-    attr_accessor :job_id, :state, :options
+    include BackgroundWorker::State
+    attr_accessor :job_id, :options
 
     def initialize(options = {})
       @options = options.symbolize_keys
       @job_id = @options[:job_id] || SecureRandom.uuid
 
-      # Store state persistently, to enable status checkups & progress reporting
-      @state = BackgroundWorker::PersistentState.new(job_id, @options)
       log("Created #{self.class}")
       log("Options are: #{@options.inspect}")
     end
 
     def perform
       raise AbstractError, 'Must be implemented in Job Class'
-    end
-
-    # Report progress...
-    def report_progress(message)
-      state.message = message
-      state.save
-    end
-
-    # Report a minor progress -- may get called a lot, so don't save it so often
-    def report_minor_progress(message)
-      state.message = message
-
-      # Only report minor events once per second
-      @last_report ||= Time.now - 2
-      time_elapsed =  Time.now - @last_report
-      return unless time_elapsed > 1
-
-      @last_report = Time.now
-      state.save
-    end
-
-    def report_successful(message = 'Finished successfully')
-      state.set_completed(message, :successful)
     end
 
     def before_perform
@@ -60,16 +37,8 @@ module BackgroundWorker
       yield self if block_given?
     end
 
-    def report_failed(message = 'Failed', detailed_message = nil)
-      state.detailed_message = detailed_message
-      state.set_completed(message, :failed)
-    end
-
     class << self
       attr_reader :queue
-      def get_state_of(job_id)
-        BackgroundWorker::PersistentState.get_state_of(job_id)
-      end
 
       # Public method to do in background...
       def perform_later(options = {})
